+++ src/main/java/com/colegio/colegio_api/service/IAlumnoService.java
@@ -20,7 +20,7 @@ public interface IAlumnoService {
                             String nuevoNombre,
                             String nuevoApellido,
                             int nuevoDni,
-                            int nuevoCuil,
+                            long nuevoCuil,
                             int nuevaEdad,
                             char nuevoGenero);
